Fixes humbug plugin so only scrooge sources are registered for source code generation.
Previously the plugin would register all files under `src_managed` with
the compiler even files that it didn't generate causing some files to be
registered twice. The Scala 2.11 compiler no longer accepts that.

This changes the plugin to it know which files it generated and it uses
`FileFunction.cached` to cache thrift code generation.