large-arbitrary-thrift: remove tail call in quasiquotes
For Humbug, we can just map over the args and splice in the resulting
list of quasiquotes; there is no need to recursively build the generator
macro.

This optimizes; the previous code was hitting JVM bytecode size limits for
thrift of ~50 fields, and overflowing stack for >100 fields. This new version
can handle the Large.thrift fiel with 1500 fields.